+++ packages/zosfiles/__tests__/__system__/methods/delete/DeleteDataset.system.test.ts
@@ -84,6 +84,23 @@ describe("Delete Dataset", () => {
             expect(response.success).toBe(true);
             expect(response.commandResponse).toContain(ZosFilesMessages.datasetDeletedSuccessfully.message);
         });
+        it("should delete a partitioned data set with tsoAccount", async () => {
+            let error;
+            let response;
+
+            try {
+                response = await Delete.dataSet(REAL_SESSION, dsname, {tsoAccount: defaultSystem.tso.account});
+                Imperative.console.info("Response: " + inspect(response));
+            } catch (err) {
+                error = err;
+                Imperative.console.info("Error: " + inspect(error));
+            }
+
+            expect(error).toBeFalsy();
+            expect(response).toBeTruthy();
+            expect(response.success).toBe(true);
+            expect(response.commandResponse).toContain(ZosFilesMessages.datasetDeletedSuccessfully.message);
+        });
     });
 
     describe("Failure scenarios", () => {