+++ packages/zosfiles/src/methods/list/List.ts
@@ -27,6 +27,7 @@ import { IFsOptions } from "./doc/IFsOptions";
 import { IZosmfListResponse } from "./doc/IZosmfListResponse";
 import { IDsmListOptions } from "./doc/IDsmListOptions";
 import { Invoke } from "../invoke/Invoke";
+import { ZosFilesUtils } from "../../utils/ZosFilesUtils";
 
 /**
  * This class holds helper functions that are used to list data sets and its members through the z/OS MF APIs
@@ -65,17 +66,17 @@ export class List {
                 params.set("start", options.start);
             }
 
-            const reqHeaders: IHeaderContent[] = [ZosmfHeaders.ACCEPT_ENCODING];
+            const reqHeaders: IHeaderContent[] = [ZosmfHeaders.ACCEPT_ENCODING].concat(ZosFilesUtils.generateTsoHeaders(options));
             if (options.attributes) {
                 reqHeaders.push(ZosmfHeaders.X_IBM_ATTRIBUTES_BASE);
             }
             if (options.maxLength) {
-                reqHeaders.push({"X-IBM-Max-Items": `${options.maxLength}`});
+                reqHeaders.push({ "X-IBM-Max-Items": `${options.maxLength}` });
             } else {
                 reqHeaders.push(ZosmfHeaders.X_IBM_MAX_ITEMS);
             }
             if (options.responseTimeout != null) {
-                reqHeaders.push({[ZosmfHeaders.X_IBM_RESPONSE_TIMEOUT]: options.responseTimeout.toString()});
+                reqHeaders.push({ [ZosmfHeaders.X_IBM_RESPONSE_TIMEOUT]: options.responseTimeout.toString() });
             }
 
             this.log.debug(`Endpoint: ${endpoint}`);
@@ -130,7 +131,7 @@ export class List {
         ImperativeExpect.toNotBeEqual(patterns.length, 0, ZosFilesMessages.missingPatterns.message);
         const zosmfResponses: IZosmfListResponse[] = [];
 
-        for(const pattern of patterns) {
+        for (const pattern of patterns) {
             const response = await List.allMembers(session, dataSetName, { pattern, maxLength: options.maxLength, start: options.start });
             zosmfResponses.push(...response.apiResponse.items);
         }
@@ -148,7 +149,7 @@ export class List {
         for (const pattern of options.excludePatterns || []) {
             const response = await List.allMembers(session, dataSetName, { pattern });
             response.apiResponse.items.forEach((membersObj: IZosmfListResponse) => {
-                const responseIndex = zosmfResponses.findIndex(response=> response.member === membersObj.member);
+                const responseIndex = zosmfResponses.findIndex(response => response.member === membersObj.member);
                 if (responseIndex !== -1) {
                     zosmfResponses.splice(responseIndex, 1);
                 }
@@ -197,17 +198,17 @@ export class List {
                 endpoint = `${endpoint}&start=${encodeURIComponent(options.start)}`;
             }
 
-            const reqHeaders: IHeaderContent[] = [ZosmfHeaders.ACCEPT_ENCODING];
+            const reqHeaders: IHeaderContent[] = [ZosmfHeaders.ACCEPT_ENCODING].concat(ZosFilesUtils.generateTsoHeaders(options));
             if (options.attributes) {
                 reqHeaders.push(ZosmfHeaders.X_IBM_ATTRIBUTES_BASE);
             }
             if (options.maxLength) {
-                reqHeaders.push({"X-IBM-Max-Items": `${options.maxLength}`});
+                reqHeaders.push({ "X-IBM-Max-Items": `${options.maxLength}` });
             } else {
                 reqHeaders.push(ZosmfHeaders.X_IBM_MAX_ITEMS);
             }
             if (options.responseTimeout != null) {
-                reqHeaders.push({[ZosmfHeaders.X_IBM_RESPONSE_TIMEOUT]: options.responseTimeout.toString()});
+                reqHeaders.push({ [ZosmfHeaders.X_IBM_RESPONSE_TIMEOUT]: options.responseTimeout.toString() });
             }
 
             // Migrated recall options
@@ -257,9 +258,9 @@ export class List {
         ImperativeExpect.toNotBeEqual(path.trim(), "", ZosFilesMessages.missingUSSFileName.message);
 
         // Error out if someone tries to use a second table parameter without specifying a first table parameter
-        if (options.depth || options.filesys != null || options.symlinks != null){
+        if (options.depth || options.filesys != null || options.symlinks != null) {
             if (!(options.group || options.user || options.name || options.size || options.mtime || options.perm || options.type)) {
-                throw new ImperativeError({msg: ZosFilesMessages.missingRequiredTableParameters.message});
+                throw new ImperativeError({ msg: ZosFilesMessages.missingRequiredTableParameters.message });
             }
         }
 
@@ -271,14 +272,14 @@ export class List {
             let endpoint = ZosFilesConstants.RESOURCE + ZosFilesConstants.RES_USS_FILES +
                 `?${ZosFilesConstants.RES_PATH}=${encodeURIComponent(path)}`;
 
-            const reqHeaders: IHeaderContent[] = [ZosmfHeaders.ACCEPT_ENCODING];
+            const reqHeaders: IHeaderContent[] = [ZosmfHeaders.ACCEPT_ENCODING].concat(ZosFilesUtils.generateTsoHeaders(options));
             if (options.maxLength) {
-                reqHeaders.push({"X-IBM-Max-Items": `${options.maxLength}`});
+                reqHeaders.push({ "X-IBM-Max-Items": `${options.maxLength}` });
             } else {
                 reqHeaders.push(ZosmfHeaders.X_IBM_MAX_ITEMS);
             }
             if (options.responseTimeout != null) {
-                reqHeaders.push({[ZosmfHeaders.X_IBM_RESPONSE_TIMEOUT]: options.responseTimeout.toString()});
+                reqHeaders.push({ [ZosmfHeaders.X_IBM_RESPONSE_TIMEOUT]: options.responseTimeout.toString() });
             }
 
             // Start modifying the endpoint with the query parameters that were passed in
@@ -332,17 +333,17 @@ export class List {
                 endpoint += `/?${ZosFilesConstants.RES_FSNAME}=${encodeURIComponent(options.fsname)}`;
             }
 
-            const reqHeaders: IHeaderContent[] = [ZosmfHeaders.ACCEPT_ENCODING];
+            const reqHeaders: IHeaderContent[] = [ZosmfHeaders.ACCEPT_ENCODING].concat(ZosFilesUtils.generateTsoHeaders(options));
             // if (options.path) {
             //     reqHeaders.push(ZosmfHeaders.X_IBM_ATTRIBUTES_BASE);
             // }
             if (options.maxLength) {
-                reqHeaders.push({"X-IBM-Max-Items": `${options.maxLength}`});
+                reqHeaders.push({ "X-IBM-Max-Items": `${options.maxLength}` });
             } else {
                 reqHeaders.push(ZosmfHeaders.X_IBM_MAX_ITEMS);
             }
             if (options.responseTimeout != null) {
-                reqHeaders.push({[ZosmfHeaders.X_IBM_RESPONSE_TIMEOUT]: options.responseTimeout.toString()});
+                reqHeaders.push({ [ZosmfHeaders.X_IBM_RESPONSE_TIMEOUT]: options.responseTimeout.toString() });
             }
 
             this.log.debug(`Endpoint: ${endpoint}`);
@@ -379,14 +380,14 @@ export class List {
                 endpoint += `/?${ZosFilesConstants.RES_PATH}=${encodeURIComponent(options.path)}`;
             }
 
-            const reqHeaders: IHeaderContent[] = [ZosmfHeaders.ACCEPT_ENCODING];
+            const reqHeaders: IHeaderContent[] = [ZosmfHeaders.ACCEPT_ENCODING].concat(ZosFilesUtils.generateTsoHeaders(options));
             if (options.maxLength) {
-                reqHeaders.push({"X-IBM-Max-Items": `${options.maxLength}`});
+                reqHeaders.push({ "X-IBM-Max-Items": `${options.maxLength}` });
             } else {
                 reqHeaders.push(ZosmfHeaders.X_IBM_MAX_ITEMS);
             }
             if (options.responseTimeout != null) {
-                reqHeaders.push({[ZosmfHeaders.X_IBM_RESPONSE_TIMEOUT]: options.responseTimeout.toString()});
+                reqHeaders.push({ [ZosmfHeaders.X_IBM_RESPONSE_TIMEOUT]: options.responseTimeout.toString() });
             }
 
             this.log.debug(`Endpoint: ${endpoint}`);