+++ packages/core/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe core SDK package will be documented in this file.
 
+## Recent Changes
+
+- Enhancement: Added `X_IBM_REQUEST_ACCTNUM` and `X_IBM_REQUEST_PROC` header constants to `ZosmfHeaders` for sending a TSO account number and logon procedure on z/OSMF data set and file requests. [#2653](https://github.com/zowe/zowe-cli/issues/2653)
+
 ## `8.35.0`
 
 - BugFix: Improved the handling of APIML service base paths in the `Services.getServicesByConfig` function. [#2819](https://github.com/zowe/zowe-cli/pull/2819)